+++ src/components/progress-ring/progress-ring.ts
@@ -1,6 +1,5 @@
 import { LitElement, html } from 'lit';
 import { customElement, property, query } from 'lit/decorators.js';
-import { watch } from '../../internal/watch';
 import styles from './progress-ring.styles';
 
 /**
@@ -12,7 +11,9 @@ import styles from './progress-ring.styles';
  * @csspart base - The component's base wrapper.
  * @csspart label - The progress ring label.
  *
- * @cssproperty --track-color - The track color.
+ * @cssproperty --size - The diameter of the progress ring (cannot be a percentage).
+ * @cssproperty --track-width - The width of the track.
+ * @cssproperty --track-color - The color of the track.
  * @cssproperty --indicator-color - The indicator color.
  */
 @customElement('sl-progress-ring')
@@ -21,28 +22,8 @@ export default class SlProgressRing extends LitElement {
 
   @query('.progress-ring__indicator') indicator: SVGCircleElement;
 
-  /** The size of the progress ring in pixels. */
-  @property({ type: Number }) size = 128;
-
-  /** The stroke width of the progress ring in pixels. */
-  @property({ attribute: 'stroke-width', type: Number }) strokeWidth = 4;
-
   /** The current progress percentage, 0 - 100. */
   @property({ type: Number, reflect: true }) percentage: number;
-
-  firstUpdated() {
-    this.updateProgress();
-  }
-
-  @watch('percentage', { waitUntilFirstUpdate: true })
-  updateProgress() {
-    const radius = this.indicator.r.baseVal.value;
-    const circumference = radius * 2 * Math.PI;
-    const offset = circumference - (this.percentage / 100) * circumference;
-
-    this.indicator.style.strokeDasharray = `${circumference} ${circumference}`;
-    this.indicator.style.strokeDashoffset = `${offset}`;
-  }
 
   render() {
     return html`
@@ -54,26 +35,9 @@ export default class SlProgressRing extends LitElement {
         aria-valuemax="100"
         aria-valuenow="${this.percentage}"
       >
-        <svg class="progress-ring__image" width=${this.size} height=${this.size}>
-          <circle
-            class="progress-ring__track"
-            stroke-width="${this.strokeWidth}"
-            stroke-linecap="round"
-            fill="transparent"
-            r=${this.size / 2 - this.strokeWidth * 2}
-            cx=${this.size / 2}
-            cy=${this.size / 2}
-          ></circle>
-
-          <circle
-            class="progress-ring__indicator"
-            stroke-width="${this.strokeWidth}"
-            stroke-linecap="round"
-            fill="transparent"
-            r=${this.size / 2 - this.strokeWidth * 2}
-            cx=${this.size / 2}
-            cy=${this.size / 2}
-          ></circle>
+        <svg class="progress-ring__image" style="--percentage: ${this.percentage}; --pi: ${Math.PI.toFixed(10)}">
+          <circle class="progress-ring__track"></circle>
+          <circle class="progress-ring__indicator"></circle>
         </svg>
 
         <span part="label" class="progress-ring__label">