+++ src/components/spinner/spinner.ts
@@ -8,9 +8,9 @@ import styles from './spinner.styles';
  *
  * @csspart base - The component's base wrapper.
  *
- * @cssproperty --track-width - The width of the indicator.
- * @cssproperty --track-color - The color of the spinner's track.
- * @cssproperty --indicator-color - The color of the spinner's indicator.
+ * @cssproperty --track-width - The width of the track.
+ * @cssproperty --track-color - The color of the track.
+ * @cssproperty --indicator-color - The color of the indicator.
  * @cssproperty --speed - The time it takes for the spinner to complete one animation cycle.
  */
 @customElement('sl-spinner')